requirements.txt
flask>=2.0
python-dateutil>=2.8
fpdf2>=2.7
openpyxl>=3.1
